fix: all endpoint work

+++ src/main/java/hei/grade/school/service/SemesterService.java
@@ -1,15 +1,17 @@
 package hei.grade.school.service;
 
+import java.time.LocalDate;
 import java.util.List;
 
 import javax.transaction.Transactional;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
-import hei.grade.school.mapper.SemesterMapper;
 import hei.grade.school.model.Semester;
 import hei.grade.school.repository.SemesterRepository;
 import lombok.AllArgsConstructor;
+import org.springframework.web.server.ResponseStatusException;
 
 @Service
 @AllArgsConstructor
@@ -21,16 +23,31 @@ public class SemesterService {
         return semesterRepository.findAll();
     }
 
-    public Semester addSemester(SemesterMapper semesterMapper) {
-        Semester semester = new Semester();
-        semester.setName(semesterMapper.getName());
-        semester.setStartDate(semesterMapper.getStartDate());
-        semester.setEndDate(semesterMapper.getEndDate());
-        semester.setStatus(semesterMapper.isStatus());
+    public Semester addSemester(Semester semester) {
+        Semester newSemester = new Semester();
+       try {
+           if(semester.getStatus()!=null){
+               newSemester.setStatus(semester.getStatus());
+           }
+           if(semester.getName()!=null){
+               newSemester.setName(semester.getName());
+           }
 
-        semesterRepository.save(semester);
+           if(semester.getStartDate()!=null){
+               String date = String.valueOf(semester.getStartDate());
+               newSemester.setStartDate(LocalDate.parse(date));
+           }
+           if(semester.getEndDate()!=null){
+               String date2 = String.valueOf(semester.getEndDate());
+               newSemester.setEndDate(LocalDate.parse(date2));
+           }
 
-        return semesterRepository.findById(semester.getId()).get();
+        semesterRepository.save(newSemester);
+       } catch ( ResponseStatusException e){
+            new ResponseStatusException(HttpStatus.BAD_REQUEST,
+                    "Server Error: Unable to update semester");
+        }
+        return semesterRepository.findById(newSemester.getId()).get();
     }
 
     public Semester findSemesterById(String id) {
@@ -38,20 +55,51 @@ public class SemesterService {
     }
 
     @Transactional
-    public Semester updateSemesterById(String id, SemesterMapper semesterMapper) {
+    public Semester updateSemesterById(String id, Semester semester) {
 
-        Semester semester = semesterRepository.findById(id).get();
-        semester.setName(semesterMapper.getName());
-        semester.setStartDate(semesterMapper.getStartDate());
-        semester.setEndDate(semesterMapper.getEndDate());
-        semester.setStatus(semesterMapper.isStatus());
+        Boolean semesterExists =  semesterRepository.existsById(id);
+        if(!semesterExists){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+                    String.format("Error: semester with id %s not found in database ", id));
+        }
 
-        semesterRepository.save(semester);
+        try {
+            Semester newSemester = semesterRepository.findById(id).get();
+            if(semester.getStatus()!=null
+                    && !semester.getStatus().equals(newSemester.getStatus())){
+                newSemester.setStatus(semester.getStatus());
+            }
+            if(semester.getName()!=null
+                    && !semester.getName().equals(newSemester.getName())){
+                newSemester.setName(semester.getName());
+            }
 
-        return semesterRepository.findById(semester.getId()).get();
+            if(semester.getStartDate()!=null
+                    && !semester.getStartDate().equals(newSemester.getStartDate())){
+                String date = String.valueOf(semester.getStartDate());
+                newSemester.setStartDate(LocalDate.parse(date));
+            }
+            if(semester.getEndDate()!=null
+                    && !semester.getEndDate().equals(newSemester.getEndDate())){
+                String date2 = String.valueOf(semester.getEndDate());
+                newSemester.setEndDate(LocalDate.parse(date2));
+            }
+
+            semesterRepository.save(newSemester);
+        } catch (ResponseStatusException e){
+            new ResponseStatusException(HttpStatus.BAD_REQUEST,
+                    "Server Error: Unable to update semester");
+        }
+
+        return semesterRepository.findById(id).get();
     }
 
     public String deleteSemesterById(String id) {
+        Boolean semesterExists =  semesterRepository.existsById(id);
+        if(!semesterExists){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+                    String.format("Error: semester with id %s not found in database ", id));
+        }
 
         semesterRepository.deleteById(id);
         return "Evaluation deleted With Success";
